Refactor code role legend highlighter
Agent-Logs-Url: https://github.com/gui-cs/Terminal.Gui/sessions/89b866c3-6b09-44f6-b2b0-e4f10515d475
