+++ Terminal.Gui/Terminal.Gui.csproj
@@ -97,6 +97,7 @@
         <InternalsVisibleTo Include="StressTests" />
         <InternalsVisibleTo Include="IntegrationTests" />
         <InternalsVisibleTo Include="TerminalGuiDesigner" />
+        <InternalsVisibleTo Include="UICatalog" />
         <InternalsVisibleTo Include="AppTestHelpers" />
         <InternalsVisibleTo Include="DynamicProxyGenAssembly2" />
     </ItemGroup>
